requirements.txt
Django>=2.1.0,<2.2.0
djangorestframework>=3.8.2,<3.9.0
flake8>=4.0.1,<4.1.0
psycopg2>=2.7.5,<2.8.0
pytest-django>=4.5.2,<5.0.0
django-extensions>=3.0.1,<4.0.0